Update Jenkinsfile
⚠️ This repository is no longer updated.

In this Bitnami Legacy repository, you can find a backup of all existing container images, which will receive no further updates or support and should only be used for temporary migration purposes. If you rely on any of these legacy images, we suggest pulling and storing them in your own container registry to ensure continued availability. Please note that this repository may be removed in the future.

For production workloads and long-term support, users are encouraged to adopt Bitnami Secure Images⁠
, which include hardened containers, smaller attack surfaces, CVE transparency (via VEX/KEV), SBOMs, and enterprise support.